refactor: reducer type & interface 파일 분리
