+++ src/advanced/components/order-summary/summary-details/ItemDetails.tsx
@@ -1,12 +1,13 @@
 import { useCartState } from '../../../contexts/CartContext';
-import { CartItemForDisplay, getCartSummary } from '../../../reducer';
+import { getCartSummary } from '../../../reducer';
+import { CartItemForDisplay } from '../../../types';
 
 export const ItemDetails = () => {
   const state = useCartState();
   const { cartItemsForDisplay } = getCartSummary(state);
 
   return cartItemsForDisplay.map((item: CartItemForDisplay) => (
-    <div className='flex justify-between text-xs tracking-wide text-gray-400'>
+    <div key={item.name} className='flex justify-between text-xs tracking-wide text-gray-400'>
       <span>
         {item.name} x {item.quantity}
       </span>
